Revert "minions-minor-modes-menu: Use x-popup-menu"
This reverts commit bc1edab09d3fc1b248c17775ae2eb7d35b96be40.

Suggested-by: Dario Gjorgjevski <dario.gjorgjevski+git@gmail.com>
  who writes:

Use popup-menu instead of x-popup-menu to make it actually work

Currently, the menu does not really do anything when an item is chosen
because x-popup-menu only returns the list of events but does not
execute the corresponding command.  From the documentation:

    When MENU is a keymap or a list of keymaps, the return value is
    the list of events corresponding to the user's choice.  Note that
    x-popup-menu does not actually execute the command bound to that
    sequence of events.

We should therefore use popup-menu instead to fix things.