Optimize mode filtering
Extract a separate function to allow compilation instead of slow
evaluation.  This optimization helps when scrolling through large
documents, where previously the minions mode line eval appeared
in the profile for redisplay.

The bottleneck is `pcase-lambda', which performs horribly when not
compiled.  Because the whole file might not be compiled drop using
it altogether.  Also use `cl-filter-if-not' instead of `seq-filter',
for an additional, though tiny speedup, and also because it allows
dropping a dependency.