minions-mode-line-lighter: Default to a hamburger menu
Previously we used ";-" as the menu button text and unless the user
customizes `minions-mode-line-delimiters' that is followed by ")" in
the mode line.  Displaying ";-)" is kinda cute at first, but it gets
old after a while, so use "≡" instead.  Hamburger menus have been
around for a long time now; so using one here should not confuse
anyone.

As a side note, personally I use

  (use-package minions
    :config (minions-mode)
    (setq minions-mode-line-lighter " ≡ ")
    (setq minions-mode-line-delimiters nil))